BMI Update

+++ Day 2/BMI calculator.py
@@ -1,4 +1,4 @@
 height = input("How tall are you? ")
 weight = input("How much do you weigh? ")
-BMI = int(weight)/(float(height)**2)
+BMI = float(weight)/(float(height)**2)
 print(BMI)
